Rename comptr file to com_ptr
